+++ tests/fixtures/fontawesome/faRemove.js
@@ -1,0 +1,15 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+var source = require('./faXmark');
+exports.definition = {
+  prefix: source.prefix,
+  iconName: source.iconName,
+  icon: [
+    source.width,
+    source.height,
+    source.aliases,
+    source.unicode,
+    source.svgPathData
+  ]};
+
+exports.faRemove = exports.definition;