+++ tests/fixtures/fontawesome/faTrashCan.js
@@ -1,0 +1,30 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+var prefix = 'fas';
+var iconName = 'trash-can';
+var width = 448;
+var height = 512;
+var aliases = [61460,"trash-alt"];
+var unicode = 'f2ed';
+var svgPathData = 'M136.7 5.9C141.1-7.2 153.3-16 167.1-16z';
+
+exports.definition = {
+  prefix: prefix,
+  iconName: iconName,
+  icon: [
+    width,
+    height,
+    aliases,
+    unicode,
+    svgPathData
+  ]};
+
+exports.faTrashCan = exports.definition;
+exports.prefix = prefix;
+exports.iconName = iconName;
+exports.width = width;
+exports.height = height;
+exports.ligatures = aliases;
+exports.unicode = unicode;
+exports.svgPathData = svgPathData;
+exports.aliases = aliases;