+++ tests/fixtures/mui/index.js
@@ -1,0 +1,12 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+Object.defineProperty(exports, "Delete", {
+  enumerable: true,
+  get: function () {
+    return _Delete.default;
+  }
+});
+var _Delete = _interopRequireDefault(require("./Delete"));