+++ tests/fixtures/remix/index.mjs
@@ -1,0 +1,5 @@
+/**
+ * @remixicon/react v4.1.0 - Remix Icon License 1.0
+ */
+
+import t from"react";const s=({color:C="currentColor",size:e=24,className:l,...i})=>t.createElement("svg",{viewBox:"0 0 24 24",xmlns:"http://www.w3.org/2000/svg",width:e,height:e,fill:C,...i,className:"remixicon "+(l||"")},t.createElement("path",{d:"M4 8H20V21H4V8ZM7 5V3H17V5H22V7H2V5H7Z"})),r=({color:C="currentColor",size:e=24,className:l,...i})=>t.createElement("svg",{viewBox:"0 0 24 24",xmlns:"http://www.w3.org/2000/svg",width:e,height:e,fill:C,...i,className:"remixicon "+(l||"")},t.createElement("path",{d:"M17 6H22V8H20V21H4V8H2V6H7V3H17V6Z"})),$a=({color:C="currentColor",size:e=24,className:l,...i})=>t.createElement("svg",{viewBox:"0 0 24 24",xmlns:"http://www.w3.org/2000/svg",width:e,height:e,fill:C,...i,className:"remixicon "+(l||"")},t.createElement("path",{d:"M8 11H12.5C13.8807 11 15 9.88071 15 8.5Z"}));export{$a as RiBold,s as RiDeleteBinFill,r as RiDeleteBinLine};